[ci-skip] Added Press Section

+++ .travis/build.py
@@ -11,19 +11,24 @@ from PIL import Image
 templateLoader = jinja2.FileSystemLoader(searchpath="./")
 templateEnv = jinja2.Environment(loader=templateLoader,trim_blocks=True)
 
+NO_DOWNLOAD = False
+
 BLOG_API_URL = "https://blog.haideralipunjabi.com/index.json"
 GITHUB_API_URL = "https://api.github.com/users/%s/repos?per_page=100&page=%s"
 FOSS_CONTRIBUTIONS_DATA = "data/foss-contributions.json"
 BLOG_IMAGE_SIZE = 512,512
+
 def get_blog_posts(num):
     data = requests.get(BLOG_API_URL).json()
-    os.system("rm -rf assets/img/blogs")
-    os.system("mkdir assets/img/blogs")
+    if not NO_DOWNLOAD:
+        os.system("rm -rf assets/img/blogs")
+        os.system("mkdir assets/img/blogs")
     for post in data[:num]:
         filein = f'assets/img/blogs/{post["data"]["image"].split("/")[-1]}'
         fileout1 = filein.rsplit(".",1)[0] + ".webp"
         fileout2 = filein.rsplit(".",1)[0] + ".jpg"
-        wget.download(post["data"]["image"],out=filein)
+        if not NO_DOWNLOAD:
+            wget.download(post["data"]["image"],out=filein)
         post["data"]["image"] = [fileout1,fileout2]
         im = Image.open(filein)
         im.thumbnail(BLOG_IMAGE_SIZE,Image.ANTIALIAS)
@@ -31,6 +36,34 @@ def get_blog_posts(num):
         im.save(fileout2)
         
     return {"blog_posts":data[:num]}
+
+def get_press_posts(num):
+    data = json.load(open("data/press.json","r"))
+    if not NO_DOWNLOAD:
+        os.system("rm -rf assets/img/press")
+        os.system("mkdir assets/img/press")
+    for i,post in enumerate(data["press"]):
+        if "title" not in post.keys() or post["title"] == "":
+            r = requests.get(f"https://api.urlmeta.org/?url={post['link']}", headers={
+                "Authorization": f"Basic {os.getenv('URLMETA_API')}"
+            }).json()
+            post["title"] = r["meta"]["title"]
+            post["description"]=r["meta"]["description"]
+            post["image_src"] = r["meta"]["image"]
+            post["site"] = r["meta"]["site"]["name"]
+        if num > i:    
+            filein = f'assets/img/press/{i}.{post["image_src"].split(".")[-1]}'
+            fileout1 = filein.rsplit(".",1)[0] + ".webp"
+            fileout2 = filein.rsplit(".",1)[0] + ".jpg"
+            if not NO_DOWNLOAD:
+                wget.download(post["image_src"],out=filein)
+            im = Image.open(filein)
+            im.thumbnail(BLOG_IMAGE_SIZE,Image.ANTIALIAS)
+            im.save(fileout1)
+            im.save(fileout2)
+            post["images"] = [fileout1,fileout2]
+    json.dump(data,fp=open("data/press.json","w"))
+    return {"press":data["press"]}
 
 def get_github_data(username,num):
     repos = []
@@ -67,7 +100,7 @@ templates = [
     {
         "input": "index.html",
         "data_files": ["backpack.json","settings.json","projects.json","timeline.json"],
-        "data":[get_blog_posts(4),get_github_data("haideralipunjabi",4),get_foss_contributions()],
+        "data":[get_blog_posts(4),get_github_data("haideralipunjabi",4),get_foss_contributions(),get_press_posts(4)],
         "output": "index.html"
     },
     {
